requirements.txt
Pillow==8.4.0
pyheif==0.5.2
